requirements.txt
streamlit==1.32.0
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.9
python-dotenv==1.0.0
groq>=0.4.0
sentence-transformers==2.6.0
chromadb==0.4.24
numpy>=1.26.0
torch>=2.2.0
transformers>=4.37.2
huggingface_hub>=0.20.3
pysqlite3-binary>=0.5.1